@marcfehling 2024-02-09
These are the relevant timings for the Stokes 3D reference scenario in the multigrid part. The machine I ran the experiments on was shared, so timings are potentially larger than in reality. For this scenario, it shows a speed-up of 65% on the setup of the smoother preconditioners, and a speed-up of 25% on total runtime.

Before:
```
Running with dealii & Trilinos on 96 MPI rank(s)...
| Section                    | no. calls |   min time  rank |   avg time |   max time  rank |
+----------------------------------------+------------------+------------+------------------+
| full_cycle                 |         1 |     91.23s    88 |      91.8s |     92.15s     9 |
| mg_setup_levels            |         1 |     41.37s     1 |     41.39s |     41.42s    67 |
|   mg_setup_level_smoothers |        13 |     39.02s    58 |     39.08s |     39.14s    30 |  <---
| mg_reinit_transfer         |         1 |    0.5896s    67 |    0.6178s |    0.6407s     1 |
| mg_solve                   |         1 |     44.07s    60 |     44.25s |     44.29s    36 |
```

After:
```
| Section                    | no. calls |   min time  rank |   avg time |   max time  rank |
+----------------------------------------+------------------+------------+------------------+
| full_cycle                 |         1 |     66.44s    23 |     67.03s |     67.35s    29 |
| mg_setup_levels            |         1 |     15.99s    26 |     16.01s |     16.04s    47 |
|   mg_setup_level_smoothers |        13 |     13.54s    13 |     13.58s |     13.65s    32 |  <---
| mg_reinit_transfer         |         1 |    0.6313s    47 |    0.6539s |    0.6738s    25 |
| mg_solve                   |         1 |     44.37s    58 |     44.54s |     44.59s    64 |
```