@marcfehling 2024-02-09
I will try this approach as well for the Poisson assembly.