@kronbichler 2024-02-10
Excellent! This looks more reasonable (assembly should rarely take more than 10% of run time in a FEM solver). If I understand the numbers correctly, `mg_setup_level_smoothers` now also contains the part where the additive Schwarz smoother factorizes its matrices as well, so the pure assembly time improved even more. I am glad this will make life slightly easier when doing the experiments.